Wrote an option to force NI weights to sum to 1
This is for the specific case of calculating single indicator indices without imputations.
The original code used BSunit weights directly, resulting in indicator indices becoming too low when only a subset of areas had data available. Setting enforce_weightSum1 = TRUE rescaled weights to sum to 1, thus returning weighted means for the areas with data that are "comparable". This has not been tested with any multi-indicator setup or with imputations.